% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcharts.R
\docType{package}
\name{clcharts}
\alias{clcharts}
\title{clcharts: Tools for making charts in the House of Commons Library style}
\description{
The clcharts package provides themes, colors, and tools for making charts in
the House of Commons Library style.
}
