% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{commonslib_colors}
\alias{commonslib_colors}
\title{Palette colors as a named vector}
\format{
An object of class \code{character} of length 10.
}
\usage{
commonslib_colors
}
\description{
Palette colors as a named vector
}
\keyword{internal}
