% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{commonslib_party_color}
\alias{commonslib_party_color}
\title{Get the hex code for a commons library party color}
\usage{
commonslib_party_color(party_name)
}
\arguments{
\item{party_name}{The name of the party.}
}
\value{
An RGB hex string for the given party color.
}
\description{
Get the hex code for a commons library party color
}
