% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{commonslib_party_colors}
\alias{commonslib_party_colors}
\title{Party colors as a named vector}
\format{An object of class \code{character} of length 16.}
\usage{
commonslib_party_colors
}
\description{
Party colors as a named vector
}
\keyword{internal}
