% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_commonslib_colors}
\alias{get_commonslib_colors}
\title{Function to extract commonslib colors as hex codes}
\usage{
get_commonslib_colors(...)
}
\arguments{
\item{...}{Names of colors in \code{commonslib_colors}.}
}
\description{
Function to extract commonslib colors as hex codes
}
\keyword{internal}
