% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{load_fonts}
\alias{load_fonts}
\title{Load the font files needed for the theme}
\usage{
load_fonts(
  national_semibold = getOption("clcharts.national_semibold"),
  national_book = getOption("clcharts.national_book"),
  open_sans_regular = getOption("clcharts.open_sans_regular"),
  open_sans_bold = getOption("clcharts.open_sans_bold")
)
}
\arguments{
\item{national_semibold}{Path to the font file for National-Semibold}

\item{national_book}{Path to the font file for National-Book}

\item{open_sans_regular}{Path to the font file for OpenSans-Regular}

\item{open_sans_bold}{Path to the font file for OpenSans-Regular}
}
\description{
This function is used to load the font files needed for the theme. It is
called automatically when the package loads. By default it looks for the
required files in a standard location: a folder in the user's home directory
called \code{.clcharts}.
}
\details{
This function is needed because there are too many inconsistencies in the
way R handles fonts on different operating systems. Rather than trying to
find and load the correct system fonts in all circustances, a font bundle is
available to package users which contains the licensed fonts.
}
\keyword{internal}
